+++ src/activationFunc.h
@@ -2,7 +2,7 @@
 #include <math.h>
 
 static struct activationFunc {
-	static std::function<double(double, uint16_t, uint16_t)> sigmoid;
+	static std::function<float(float, uint16_t, uint16_t)> sigmoid;
 
-	static std::function<double(double, uint16_t, uint16_t)> dsigmoid; 
+	static std::function<float(float, uint16_t, uint16_t)> dsigmoid; 
 };
