+++ src/matrix.h
@@ -10,29 +10,29 @@
 class matrix {
     private:
         uint16_t rows,cols;
-        double **data;
+        float **data;
     public:
         matrix(uint16_t rows,uint16_t cols);
         ~matrix();
 
-        static matrix* fromArray(std::vector<double> arr);
+        static matrix* fromArray(std::vector<float> arr);
         static void randomize(matrix* m);
-		static std::vector<double> toArray(matrix* m);
+		static std::vector<float> toArray(matrix* m);
 		static matrix* dotProduct(matrix* m1, matrix* m2);
-		static matrix* map(matrix* m, std::function<double(double, uint16_t, uint16_t)> func);
+		static matrix* map(matrix* m, std::function<float(float, uint16_t, uint16_t)> func);
 		static matrix* transpose(matrix* m);
 
         matrix operator+ (const matrix& b);
 		matrix operator+= (const matrix& b);
         matrix operator- (const matrix& b);
         matrix operator* (const matrix& b);
-        matrix operator* (const double b);
+        matrix operator* (const float b);
 
 		matrix* clone();
 
-		void map(std::function<double(double, uint16_t, uint16_t)> func);
+		void map(std::function<float(float, uint16_t, uint16_t)> func);
         void print();
 		void transpose();
 
-		double** getMatrixArr() const;
+		float** getMatrixArr() const;
 };
